refactor: 서비스 계층에서는 http와 무관한 비즈니스 로직만 담당하도록 변경
