+++ src/main/java/com/sprint/mission/discodeit/service/AuthService.java
@@ -3,12 +3,11 @@ package com.sprint.mission.discodeit.service;
 import com.sprint.mission.discodeit.dto.auth.UserRoleUpdateRequest;
 import com.sprint.mission.discodeit.dto.user.response.UserResponseDto;
 import com.sprint.mission.discodeit.security.jwt.JwtDto;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.data.util.Pair;
 
 public interface AuthService {
 
     UserResponseDto updateRole(UserRoleUpdateRequest request);
 
-    JwtDto refreshAccessToken(HttpServletRequest request, HttpServletResponse response);
+    Pair<JwtDto, String> refreshAccessToken(String refreshToken);
 }
